flask-server-back-end/requirements.txt
Flask==2.0.3
boto3==1.21.12
python-decouple==3.6
pandas==1.4.1
requests==2.27.1
Flask-Cors==3.0.10 
pytest==7.1.1